tools/uploader/pyproject.toml
[project]
name = "magic-compare-uploader"
version = "0.1.0"
description = "Local CLI importer for Magic Compare Web"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.28.1",
  "pillow>=11.2.1",
  "pyyaml>=6.0.2",
  "rich>=14.1.0",
  "typer>=0.15.4"
]

[project.scripts]
magic-compare-uploader = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
